+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "javascript.validate.enable": false
+    "typescript.validate.enable": false
 }
