+++ backend/controllers/fundingitems.controller.js
@@ -2,109 +2,7 @@ const FundingItem = require('../models/fundingitem.model')
 const SponsorshipFund = require('../models/sponsorshipfund.model')
 const PersonalPurchase = require('../models/personalpurchase.model')
 const UWFinancePurchase = require('../models/uwfinancepurchase.model')
-const Constants = require('../models/constants')
-
-// empty list arg queries for all Funding Items
-const getAnnotatedFundingItemsByIdList = async (idList = []) => {
-    if (idList.length === 0) {
-        idList = await FundingItem.distinct('_id')
-    }
-    const fundingItemList = await Promise.all(
-        idList.map(async (id) =>
-            FundingItem.aggregate([
-                {
-                    $match: {
-                        _id: parseInt(id),
-                    },
-                },
-                // load ppr and upr documents as temp variables via links to referenced collection
-                {
-                    $lookup: {
-                        from: 'personalpurchases',
-                        localField: 'ppr_links',
-                        foreignField: '_id',
-                        as: 'ppr_docs',
-                    },
-                },
-                {
-                    $lookup: {
-                        from: 'uwfinancepurchases',
-                        localField: 'upr_links',
-                        foreignField: '_id',
-                        as: 'upr_docs',
-                    },
-                },
-                // filter documents by status (only include docs that count towards funding_spent)
-                {
-                    $set: {
-                        ppr_docs: {
-                            $filter: {
-                                input: '$ppr_docs',
-                                as: 'ppr_doc',
-                                cond: {
-                                    $in: [
-                                        '$$ppr_doc.status',
-                                        Constants.PPR_STATUS_FUNDING_SPENT,
-                                    ],
-                                },
-                            },
-                        },
-                    },
-                },
-                {
-                    $set: {
-                        upr_docs: {
-                            $filter: {
-                                input: '$upr_docs',
-                                as: 'upr_doc',
-                                cond: {
-                                    $in: [
-                                        '$$upr_doc.status',
-                                        Constants.UPR_STATUS_FUNDING_SPENT,
-                                    ],
-                                },
-                            },
-                        },
-                    },
-                },
-                {
-                    $set: {
-                        type: 'FI',
-                        code: {
-                            $concat: ['FI-', { $toString: '$_id' }],
-                        },
-                        path: {
-                            $concat: ['/FI/', { $toString: '$_id' }],
-                        },
-                        funding_spent: {
-                            $round: [
-                                {
-                                    $sum: [
-                                        {
-                                            $sum: '$ppr_docs.cost',
-                                        },
-                                        {
-                                            $sum: '$upr_docs.cost',
-                                        },
-                                    ],
-                                },
-                                2,
-                            ],
-                        },
-                    },
-                },
-                {
-                    $unset: ['ppr_docs', 'upr_docs'],
-                },
-            ])
-        )
-    )
-
-    // aggregate returns an array so fundingItemList
-    // will always be a list of one-elem lists:
-    // i.e. [[FI-1], [FI-2], ...] where FI-X is a FundingItem object
-    return fundingItemList.flat()
-}
+const { getAnnotatedFundingItemsByIdList } = require('./annotatedGetters')
 
 const getAllFundingItems = (_, res) => {
     getAnnotatedFundingItemsByIdList()
